Trying to do PB
These changes allow the user to drag and drop the flag anywhere on the webpage. However, it's flawed and allows the user to drag it even out of the border. Moreover, the flag icon seems to be distorted. While the basic functionality is there, more refinement is needed. This also does not show the progress bar, have the backspace, or trash.